JDF-366 Upgrade RequireJS to 2.1.8
Upgraded the RequireJS version to prevent occassional failures in MSIE 10. Removed the order plugin since it is no longer required.

Reworked the configuration of jQuery Mobile to disable the jQuery Mobile router on mobileinit. The router is also defined and initialized only after the binding on mobileinit is established, since the router module has a dependency on jQuery Mobile.

Also updated the RequireJS text plugin to 2.0.8.