Updated tutorial for JBDS 8.
Corrected mistakes in the tutorial.

Ordered the entities to be written by the reader, in a more
logical manner.

Corrected code listings of JPA entities. Corrected the Data
Persistence chapter to include missing entities.

Corrected the code listing for the BookingServiceTest in the
Business Logic chapter. Added a tip linking the Arquillian guides.

Corrected the chapter on building the frontend.

Also updated the instructions in the Introduction for JBDS 8.

Corrected instructions on adding Jackson Core from RESTEasy , and
ShrinkWrap resolvers from WFK Tools BOM.

Corrected JBoss BOM versions used in the documentation.

Corrected Forge usage in JBDS 8.